requirements.txt
hydra-core
omegaconf
tqdm
h5py
scipy
scikit-learn
tensorboardX